Generic Solver
Many moons ago, this project was simply about manufacturing a DFA for an arbitrary language defined by a function. I got away form that, and by the time I was fluent enough in Rust to respect that initial goal, everything had long been hardcoded to work in the domain of SRSs exclusively. This update restores the inherently generic ability of the BFSSolver (but I still need to do so for HashSolver), alongside restructuring everything along a more generic infrastructure that should make future work and modification much easier.